requirements.txt
flask-sqlalchemy
flask
requests
flask-crontab
